Allow setting VERIFY_IDENTITY for MariaDB
This adds support for setting the VERIFY_IDENTITY mode with MariaDB. On
MariaDB, the `MYSQL_OPT_SSL_VERIFY_SERVER_CERT` option is available
which is equivalent to `VERIFY_IDENTITY`.

Many containers with Ruby apps are based on Debian where MariaDB is the
standard provided, so this improves support there significantly.